task: add gzip/unzip support to tracker/events
