+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/tracker/export/csv/TrackerCsvTrackedEntityService.java
@@ -39,6 +39,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 import org.hisp.dhis.dxf2.events.event.csv.CsvEventService;
+import org.hisp.dhis.webapi.controller.tracker.export.CompressionUtil;
 import org.hisp.dhis.webapi.controller.tracker.view.Attribute;
 import org.hisp.dhis.webapi.controller.tracker.view.TrackedEntity;
 import org.springframework.stereotype.Service;
@@ -52,13 +53,7 @@ public class TrackerCsvTrackedEntityService implements CsvEventService<TrackedEn
   public void writeEvents(
       OutputStream outputStream, List<TrackedEntity> trackedEntities, boolean withHeader)
       throws IOException {
-    final CsvSchema csvSchema =
-        CSV_MAPPER
-            .schemaFor(CsvTrackedEntity.class)
-            .withLineSeparator("\n")
-            .withUseHeader(withHeader);
-
-    ObjectWriter writer = CSV_MAPPER.writer(csvSchema.withUseHeader(withHeader));
+    ObjectWriter writer = getObjectWriter(withHeader);
 
     List<CsvTrackedEntity> attributes = new ArrayList<>();
 
@@ -97,6 +92,30 @@ public class TrackerCsvTrackedEntityService implements CsvEventService<TrackedEn
     }
 
     writer.writeValue(outputStream, attributes);
+  }
+
+  @Override
+  public void writeZip(
+      OutputStream outputStream, List<TrackedEntity> toCompress, boolean withHeader, String file)
+      throws IOException {
+    CompressionUtil.writeZip(outputStream, toCompress, getObjectWriter(withHeader), file);
+  }
+
+  @Override
+  public void writeGzip(
+      OutputStream outputStream, List<TrackedEntity> toCompress, boolean withHeader)
+      throws IOException {
+    CompressionUtil.writeGzip(outputStream, toCompress, getObjectWriter(withHeader));
+  }
+
+  private ObjectWriter getObjectWriter(boolean withHeader) {
+    final CsvSchema csvSchema =
+        CSV_MAPPER
+            .schemaFor(CsvTrackedEntity.class)
+            .withLineSeparator("\n")
+            .withUseHeader(withHeader);
+
+    return CSV_MAPPER.writer(csvSchema.withUseHeader(withHeader));
   }
 
   private String checkForNull(Instant instant) {
